move parsing out of qemu into generic module
