+++ rebg/src/tracer/mod.rs
@@ -1,6 +1,9 @@
 use crate::{arch::Arch, state::Step};
-use std::{collections::HashMap, marker::PhantomData, path::Path, process::Child};
+use std::{collections::HashMap, fmt, io::Read, marker::PhantomData, path::Path, process::Child};
 
+use self::parser::Message;
+
+pub mod parser;
 pub mod qemu;
 
 #[derive(Debug)]
@@ -33,4 +36,57 @@ where
     pub args: Vec<String>,
     // for trait inferance
     _step: PhantomData<STEP>,
+}
+
+pub fn get_next_step<R: Read, STEP, const N: usize>(
+    reader: &mut R,
+    proc: &mut Option<Child>,
+) -> Option<ParsedStep<STEP, N>>
+where
+    STEP: Step<N> + Send + 'static + fmt::Debug,
+    STEP: for<'a> TryFrom<&'a [Message], Error = anyhow::Error>,
+{
+    #[allow(clippy::question_mark)]
+    if proc.is_none() {
+        return None;
+    }
+
+    let mut msgs = vec![];
+
+    while let Some(m) = parser::get_next_message(reader) {
+        if matches!(m, Message::Separator) {
+            break;
+        }
+
+        msgs.push(m);
+    }
+
+    // if there are no msgs, we're done!
+    if msgs.is_empty() {
+        let mut my_proc = None;
+        std::mem::swap(proc, &mut my_proc);
+        let my_proc = my_proc.unwrap();
+
+        // make sure it closed gracefully
+        let result = my_proc.wait_with_output().unwrap();
+
+        return Some(ParsedStep::Final(result));
+    }
+
+    if matches!(msgs[0], Message::LibLoad(_, _, _)) {
+        let map = msgs
+            .into_iter()
+            .map(|m| match m {
+                Message::LibLoad(name, from, to) => (name.to_string(), (from, to)),
+                _ => panic!("Got libload and some other junk!"),
+            })
+            .collect();
+
+        return Some(ParsedStep::LibLoad(map));
+    }
+
+    // otherwise, it's just a step :)
+
+    let s = STEP::try_from(&msgs).unwrap();
+    Some(ParsedStep::TraceStep(s))
 }
